enhanced_classic_rag/pyproject.toml
[project]
name = "enhanced-classic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "pinecone>=8.0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.30.0",
]
